implement npx networks command
