+++ .gitignore
@@ -62,3 +62,4 @@ typings/
 .editorconfig
 .env
 .idea
+.vscode
