+++ package.json
@@ -3,6 +3,9 @@
   "description": "Hardhat plugin for integration with Tenderly",
   "version": "1.1.4",
   "main": "dist/index.js",
+  "bin": {
+    "tenderly": "dist/internal/cli/index.js"
+  },
   "types": "dist/index.d.ts",
   "repository": {
     "type": "git",
@@ -38,7 +41,7 @@
     "chai": "^4.2.0",
     "dotenv": "^6.2.0",
     "hardhat": "^2.9.9",
-    "mocha": "^5.2.0",
+    "mocha": "^10.0.0",
     "prettier": "^1.17.0",
     "source-map-support": "^0.5.12",
     "ts-node": "^8.1.0",
@@ -53,9 +56,12 @@
   "dependencies": {
     "@nomiclabs/hardhat-ethers": "^2.0.6",
     "axios": "^0.21.1",
+    "cli-table": "^0.3.11",
+    "commander": "^9.4.0",
     "ethers": "^5.6.8",
     "fs-extra": "^9.0.1",
     "hardhat-deploy": "^0.11.10",
-    "js-yaml": "^3.14.0"
+    "js-yaml": "^3.14.0",
+    "promptly": "^3.2.0"
   }
 }
