+++ src/internal/cli/commands/index.ts
@@ -1,0 +1,3 @@
+import { NetworksCommand } from "./NetworksCommand";
+
+export { NetworksCommand };
