+++ src/utils/util.ts
@@ -2,7 +2,7 @@ import { HardhatPluginError } from "hardhat/plugins";
 import { HardhatConfig } from "hardhat/src/types/config";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { PluginName } from "../index";
+import { PluginName } from "../constants";
 import {
   ContractByName,
   Metadata,
